requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-bcrypt==1.0.1
flask-jwt-extended==4.5.3
flask-cors==4.0.0
pymysql==1.1.0
requests==2.31.0
inference-sdk==0.9.15
PyJWT==2.8.0
python-dotenv==1.0.1
cryptography==42.0.5
Pillow==10.2.0
flask-socketio==5.3.6
eventlet==0.35.2
gunicorn==21.2.0
